data/requirements.txt
connexion==1.5.3
Django
django-admin-bootstrapped
django-bootstrap3
django-extensions
django-graphos
djangorestframework
django_extensions
MySQL-python
django_tables2
django-crispy-forms
Flask==1.0.2
Jinja2==2.10
jsonschema==2.6.0
pathlib==1.0.1
pbr==4.3.0
post==1.0.2
public==1.0.3
query-string==1.0.2
request==1.0.2
django-filter~=1.1
django-filters==0.2.1